requirements-test.txt
django-otp
python-u2flib-host
python-u2flib-server
coverage
factory-boy
pytest
pytest-cov
pytest-django
